Split DefaultContextResolver in 2 beans
This refactoring is a no-op in terms of functionality.

This is required as `RealmContext` and `CallContext` may be treated
differently in the future (e.g. different scopes).
