+++ dropwizard/service/src/test/java/org/apache/polaris/service/admin/PolarisOverlappingCatalogTest.java
@@ -18,7 +18,7 @@
  */
 package org.apache.polaris.service.admin;
 
-import static org.apache.polaris.service.context.DefaultContextResolver.REALM_PROPERTY_KEY;
+import static org.apache.polaris.service.context.DefaultRealmContextResolver.REALM_PROPERTY_KEY;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.dropwizard.testing.ConfigOverride;
