+++ dropwizard/service/src/test/java/org/apache/polaris/service/admin/PolarisOverlappingTableTest.java
@@ -19,7 +19,7 @@
 package org.apache.polaris.service.admin;
 
 import static org.apache.polaris.service.admin.PolarisAuthzTestBase.SCHEMA;
-import static org.apache.polaris.service.context.DefaultContextResolver.REALM_PROPERTY_KEY;
+import static org.apache.polaris.service.context.DefaultRealmContextResolver.REALM_PROPERTY_KEY;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.dropwizard.testing.ConfigOverride;
